refactor: props
