+++ components/collection/CollectionTrades.vue
@@ -12,15 +12,20 @@
         :key="key"
         :query="tradeQuery"
         :type="tradeType"
-        :enable-collection-swap="tradeType === TradeType.SWAP"
-      />
+      >
+        <template #action>
+          <CreateCollectionSwapButton v-if="isTradeSwap(tradeType)" />
+        </template>
+      </TradeActivityTable>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
 import { type TradeTableQuery } from '@/components/trade/TradeActivityTable.vue'
-import { TradeType } from '@/components/trade/types'
+import type { TradeType } from '@/components/trade/types'
+import { isTradeSwap } from '@/composables/useTradeType'
+import CreateCollectionSwapButton from '@/components/swap/CreateCollectionSwapButton.vue'
 
 defineProps<{
   tradeType: TradeType
