+++ components/trade/TradeActivityTable.vue
@@ -22,7 +22,7 @@
           </div>
         </NeoButton>
       </div>
-      <CreateCollectionSwapButton v-if="enableCollectionSwap" />
+      <slot name="action" />
     </div>
 
     <hr class="mb-10 mt-0">
@@ -101,7 +101,6 @@
 <script lang="ts" setup>
 import { NeoButton } from '@kodadot1/brick'
 import type { TradeType, TradeNftItem } from '@/components/trade/types'
-import CreateCollectionSwapButton from '@/components/swap/CreateCollectionSwapButton.vue'
 
 type TradeTabType = 'outgoing' | 'incoming'
 export type TradeTableQuery = Record<TradeTabType, string>
@@ -109,7 +108,6 @@ export type TradeTableQuery = Record<TradeTabType, string>
 const props = defineProps<{
   type: TradeType
   query: TradeTableQuery
-  enableCollectionSwap?: boolean
 }>()
 
 const route = useRoute()
